+++ CHANGELOG.md
@@ -50,6 +50,10 @@ and this project adheres to
 - ♿️(frontend) focus side panel container on open #1452
 - 🐛(summary) whisper call error handling
 
+### Added
+
+- ✨(frontend) add screen share zoom controls #1498
+
 ## [1.23.0] - 2026-07-08
 
 ### Added